+++ frontend/src/main.ts
@@ -1,9 +1,13 @@
 import { mount } from 'svelte'
 import './app.css'
 import App from './App.svelte'
+import Admin from './Admin.svelte'
 
-const app = mount(App, {
-  target: document.getElementById('app')!,
+const target = document.getElementById('app')!
+const Component = window.location.pathname === '/admin' ? Admin : App
+
+const app = mount(Component, {
+  target,
 })
 
 export default app
